New shader - toon fireball
When applied to a sphere or other rounded object, this shader warps the vertices and lerps colours to produce a pleasing, crackly fire effect. As well as being able to customise the colour, the flame direction can be set to either world or local space, allowing the flame to shift in response to any external stimuli I might decide to program in.