Noise wind billow improvements – cycling without scrolling, and switching between gradient and Voronoi noise
I added a ‘general agitation’ value, that allows the noise to shift over time without changing position. For Voronoi noise, this updates the angle offset, and for gradient noise it changes the fourth value on a new 4D noise shader (which I'll need to improve on in the future, but works well for now). This allows the material to animate from general forces that aren’t occurring in a specific direction.