Water body shader
This shader uses a variety of depth checks and noise to create a material for a body of water, including waves/rippling, bobbing, foam, opacity increasing with depth, and tinting near the horizon. The effect is entirely world-space based and ignores UVs, giving it a consistent effect no matter what mesh it's put on. Makes prominent use of a depth fade test shader that I accidentally included in an earlier, unrelated commit.

Inspired by this tutorial: https://ameye.dev/notes/stylized-water-shader/

This shader has some room for improvement, like proper waves, but I'm real pleased with it for now.