WaterLayer script that automatically generates a large mesh to allow for better waves
My water shader is mostly designed to only use world space data and be unaffected by the mesh it's put on, but one exception is the number of vertices in the mesh, which affects how much it can be deformed.
I fixed this with a script that automatically generates a flat plane of a specified resolution, with more vertices to be modified and more obvious bobbing and waves. The dimensions and density can be easily updated in the Inspector.